+++ packages/web/src/icons/__stories__/Icon.stories.tsx
@@ -3,8 +3,6 @@
  * Generated from yarn nx run icons:generate-stories
  */
 
-import React from 'react';
-
 import { Icon } from '../Icon';
 
 import { IconSheet } from './IconSheet';