+++ packages/web/src/illustrations/__stories__/HeroSquare.stories.tsx
@@ -3,8 +3,6 @@
  * Generated from yarn nx run illustrations:generate-stories
  */
 
-import React from 'react';
-
 import { HeroSquare } from '../HeroSquare';
 
 import { getIllustrationSheet } from './getIllustrationSheet';